Move ProviderErrorCode to _provider
